Support feedback with LED and add resilience
This commit adds feedback through the LED (currently hardcoded) for the
following cases:

1. When the service starts, flash at 2Hz, and stay on
2. When the button is pressed, flash at 5Hz, and stay on
3. When the service gracefully stops, flash from 5Hz to 30Hz, and stay
off

With this, there is feedback for the user. Additionally, resilience has
been added so that if any thread dies, all other are instructed to exit
and give the feedback in case (3), along with terminating the entire
application.

The button mentioned in case (2) currently performs no action, and for
now this is hardcoded as well.

The button will remain on during normal non-interactive operation, and
this was designed as to avoid any distraction while driving. For
instance, an occassional flash on file rotation or similar event was
considered, but the concern was that this may serve to distract.